requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-32
async-generator=1.10=pypi_0
attrs=22.1.0=pypi_0
ca-certificates=2022.4.26=h9f7ea03_0
certifi=2022.6.15=py38h9f7ea03_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
contourpy=1.0.6=pypi_0
cycler=0.11.0=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.0.4=pypi_0
fonttools=4.38.0=pypi_0
future=0.18.2=pypi_0
h11=0.14.0=pypi_0
idna=3.4=pypi_0
kiwisolver=1.4.4=pypi_0
lxml=4.9.1=pypi_0
matplotlib=3.6.2=pypi_0
mouseinfo=0.1.3=pypi_0
numexpr=2.8.4=pypi_0
numpy=1.23.5=pypi_0
openpyxl=3.0.10=pypi_0
openssl=1.1.1o=hc431981_0
outcome=1.2.0=pypi_0
packaging=21.3=pypi_0
pandas=1.5.2=pypi_0
pandastable=0.13.0=pypi_0
pillow=9.3.0=pypi_0
pip=21.2.4=py38h9f7ea03_0
pyautogui=0.9.53=pypi_0
pycparser=2.21=pypi_0
pygetwindow=0.0.9=pypi_0
pymsgbox=1.0.9=pypi_0
pyodbc=4.0.35=pypi_0
pyparsing=3.0.9=pypi_0
pyperclip=1.8.2=pypi_0
pyrect=0.2.0=pypi_0
pyscreeze=0.1.28=pypi_0
pysocks=1.7.1=pypi_0
python=3.8.10=hf3be9b6_7
python-dateutil=2.8.2=pypi_0
python-docx=0.8.11=pypi_0
pytweening=1.0.4=pypi_0
pytz=2022.6=pypi_0
pywin32=305=pypi_0
requests=2.28.1=pypi_0
selenium=4.7.2=pypi_0
setuptools=61.2.0=py38h9f7ea03_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlite=3.38.3=hc431981_0
trio=0.22.0=pypi_0
trio-websocket=0.9.2=pypi_0
urllib3=1.26.13=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38h9f7ea03_2
wsproto=1.2.0=pypi_0
xlrd=2.0.1=pypi_0
